+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
 ## [Unreleased]
-Nothing yet.
+### Added
+- [#8](https://github.com/OldSneerJaw/couchster/issues/8): Regular expression pattern constraint for document ID
 
 ## [0.2.0] - 2018-03-08
 ### Added